Factor out Kore.Internal.TermLike.Renaming
